+++ tests/containers/aboutTheData/UnlinkedDataContainer-test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from 'test-utils';
-import * as aboutTheDataHelper from 'helpers/aboutTheDataHelper';
+import * as agencyReportingAPI from 'apis/agencyReporting';
 import UnlinkedDataContainer from 'containers/aboutTheData/modals/UnlinkedDataContainer';
 
 const defaultProps = {
@@ -13,7 +13,7 @@ const defaultProps = {
 
 describe('Unlinked Data Container', () => {
     it('should call api one time on mount', () => {
-        const unlinkedDataRequest = jest.spyOn(aboutTheDataHelper, 'fetchUnlinkedData');
+        const unlinkedDataRequest = jest.spyOn(agencyReportingAPI, 'fetchUnlinkedData');
         render(<UnlinkedDataContainer {...defaultProps} />);
         expect(unlinkedDataRequest).toHaveBeenCalledTimes(1);
     });
